requirements.txt
Flask>=1.0
gunicorn>=19.9
psycopg2-binary>=2.7

Flask-WTF>=0.14.2
Flask-Bcrypt>=0.7.1

inflection>=0.3.1
